Consolidating similar stories files.
